Fix for APIError response properties not added to store on web...
The tranformError used to strip values that cannot be converted with fromJS uses JSON.stringify which only sees own properties. For some reason on web the properties in APIError.response are not own props on web though they are in react-native. This ensures the keys needed will always be in the store after processing.